# Check the symlink itself, not the target of the symlink. Fixes issue #52