@andrewgregory 2022-05-26
Edited to print a warning on failures other than ENOENT and merged manually.